Split asymmetric averaging and bilateral filtering
